Clean up and comments.

+++ src/examples/SyncAccessHandleVFS.js
@@ -1,8 +1,10 @@
 // Copyright 2022 Roy T. Hashimoto. All Rights Reserved.
 import * as VFS from '../VFS.js';
 
-const BLOCK_SIZE = 4096;
+const SECTOR_SIZE = 4096;
 
+// Each OPFS file begins with a fixed-size header with metadata. The
+// contents of the file follow immediately after the header.
 const HEADER_MAX_PATH_SIZE = 512;
 const HEADER_DIGEST_SIZE = 8;
 const HEADER_OFFSET_PATH = 0;
@@ -13,18 +15,36 @@ function log(...args) {
   console.debug(...args);
 }
 
+/**
+ * This VFS uses the updated Access Handle API with all synchronous methods
+ * on FileSystemSyncAccessHandle (instead of just read and write). It will
+ * work with the regular SQLite WebAssembly build, i.e. the one without
+ * Asyncify.
+ */
 export class SyncAccessHandleVFS extends VFS.Base {
   #ready;
+
+  // All the OPFS files the VFS uses are contained in one flat directory
+  // specified in the constructor. No other files should be written here.
+  #directoryPath;
   #directoryHandle;
 
-  #mapPathToAccessHandle = new Map();
+  // The OPFS files all have randomly-generated names that do not match
+  // the SQLite files whose data they contain. This map links those names
+  // with their respective OPFS access handles. In this map, all the OPFS
+  // files that are not yet associated with a SQLite file precede the
+  // OPFS files that are associated with a SQLite file.
   #mapAccessHandleToName = new Map();
+
+  // When a SQLite file is associated with an OPFS file, that association
+  // is kept in this map.
+  #mapPathToAccessHandle = new Map();
 
   #mapIdToFile = new Map();
 
   constructor(directoryPath) {
     super();
-    this.#ready = this.#initialize(directoryPath);
+    this.#directoryPath = directoryPath;
   }
 
   get name() { return 'sync-access-handle'; }
@@ -32,13 +52,17 @@ export class SyncAccessHandleVFS extends VFS.Base {
   xOpen(name, fileId, flags, pOutFlags) {
     log(`xOpen ${name} ${fileId} 0x${flags.toString(16)}`);
     try {
+      // First try to open a path that already exists in the file system.
       const path = name ? this.#getPath(name) : Math.random().toString(36);
       let accessHandle = this.#mapPathToAccessHandle.get(path);
       if (!accessHandle && (flags & VFS.SQLITE_OPEN_CREATE)) {
+        // File not found so try to create it.
         if (this.getSize() < this.getCapacity()) {
+          // Choose an unassociated OPFS file from the pool.
           ([accessHandle] = this.#mapAccessHandleToName.keys());
           this.#setAssociatedPath(accessHandle, path);
         } else {
+          // Out of unassociated files. This can be fixed with addCapacity().
           throw new Error('cannot create file');
         }
       }
@@ -47,6 +71,8 @@ export class SyncAccessHandleVFS extends VFS.Base {
       }
       this.#mapPathToAccessHandle.set(path, accessHandle);
 
+      // Subsequent methods are only passed the fileId, so make sure we have
+      // a way to get the file resources.
       const file = {
         path,
         flags,
@@ -64,13 +90,14 @@ export class SyncAccessHandleVFS extends VFS.Base {
 
   xClose(fileId) {
     const file = this.#mapIdToFile.get(fileId);
-    log(`xClose ${file.filename}`);
+    if (file) {
+      log(`xClose ${file.path}`);
 
-    this.#mapIdToFile.delete(fileId);
-    if (file.flags & VFS.SQLITE_OPEN_DELETEONCLOSE) {
-      this.#deletePath(file.path);
+      this.#mapIdToFile.delete(fileId);
+      if (file.flags & VFS.SQLITE_OPEN_DELETEONCLOSE) {
+        this.#deletePath(file.path);
+      }
     }
-
     return VFS.SQLITE_OK;
   }
 
@@ -104,7 +131,7 @@ export class SyncAccessHandleVFS extends VFS.Base {
 
   xSync(fileId, flags) {
     const file = this.#mapIdToFile.get(fileId);
-    log(`xSync ${file.filename} ${flags}`);
+    log(`xSync ${file.path} ${flags}`);
 
     file.accessHandle.flush();
     return VFS.SQLITE_OK;
@@ -112,23 +139,21 @@ export class SyncAccessHandleVFS extends VFS.Base {
 
   xFileSize(fileId, pSize64) {
     const file = this.#mapIdToFile.get(fileId);
-    log(`xFileSize ${file.path}`);
 
     const size = file.accessHandle.getSize() - HEADER_OFFSET_DATA;
+    log(`xFileSize ${file.path} ${size}`);
     pSize64.set(size);
     return VFS.SQLITE_OK;
   }
 
   xSectorSize(fileId) {
-    log('xSectorSize', BLOCK_SIZE);
-    return BLOCK_SIZE;
+    log('xSectorSize', SECTOR_SIZE);
+    return SECTOR_SIZE;
   }
 
   xDeviceCharacteristics(fileId) {
     log('xDeviceCharacteristics');
-    return VFS.SQLITE_IOCAP_SAFE_APPEND |
-           VFS.SQLITE_IOCAP_SEQUENTIAL |
-           VFS.SQLITE_IOCAP_UNDELETABLE_WHEN_OPEN;
+    return VFS.SQLITE_IOCAP_UNDELETABLE_WHEN_OPEN;
   }
 
   xAccess(name, flags, pResOut) {
@@ -153,18 +178,46 @@ export class SyncAccessHandleVFS extends VFS.Base {
     await this.#releaseAccessHandles();
   }
 
-  ready() {
-    return this.#ready;
+  /**
+   * Release and reacquire all OPFS access handles. This must be called
+   * and awaited before any SQLite call that uses the VFS and also before
+   * any capacity changes.
+   */
+  async reset() {
+    // All files are stored in a single directory.
+    let handle = await navigator.storage.getDirectory();
+    for (const d of this.#directoryPath.split('/')) {
+      if (d) {
+        handle = await handle.getDirectoryHandle(d, { create: true });
+      }
+    }
+    this.#directoryHandle = handle;
+
+    await this.#releaseAccessHandles();
+    await this.#acquireAccessHandles();
   }
 
+  /**
+   * Returns the number of SQLite files in the file system.
+   * @returns {number}
+   */
   getSize() {
     return this.#mapPathToAccessHandle.size;
   }
 
+  /**
+   * Returns the maximum number of SQLite files the file system can hold.
+   * @returns {number}
+   */
   getCapacity() {
     return this.#mapAccessHandleToName.size;
   }
 
+  /**
+   * Increase the capacity of the file system by n.
+   * @param {number} n 
+   * @returns {Promise<number>} 
+   */
   async addCapacity(n) {
     /** @type {[any, string][]} */ const newEntries = [];
     for (let i = 0; i < n; ++i) {
@@ -178,12 +231,20 @@ export class SyncAccessHandleVFS extends VFS.Base {
 
     // Insert new entries at the front of #mapAccessHandleToName.
     this.#mapAccessHandleToName = new Map([...newEntries, ...this.#mapAccessHandleToName]);
+    return n;
   }
 
+  /**
+   * Decrease the capacity of the file system by n. The capacity cannot be
+   * decreased to fewer than the current number of SQLite files in the
+   * file system.
+   * @param {number} n 
+   * @returns {Promise<number>}
+   */
   async removeCapacity(n) {
     let nRemoved = 0;
     for (const [accessHandle, name] of this.#mapAccessHandleToName) {
-      if (nRemoved == n || this.getSize() === this.getCapacity()) return;
+      if (nRemoved == n || this.getSize() === this.getCapacity()) return nRemoved;
 
       await accessHandle.close();
       await this.#directoryHandle.removeEntry(name);
@@ -192,23 +253,11 @@ export class SyncAccessHandleVFS extends VFS.Base {
     }
   }
 
-  async #initialize(directoryPath) {
-    // All files are stored in a single directory.
-    let handle = await navigator.storage.getDirectory();
-    for (const d of directoryPath.split('/')) {
-      if (d) {
-        handle = await handle.getDirectoryHandle(d, { create: true });
-      }
-    }
-    this.#directoryHandle = handle;
-
-    await this.#acquireAccessHandles();
-  }
-
   async #acquireAccessHandles() {
     /** @type {[any, string][]} */ const tuplesWithPath = [];
     /** @type {[any, string][]} */ const tuplesWithoutPath = [];
 
+    // Enumerate all the files in the directory.
     // @ts-ignore
     for await (const [name, handle] of this.#directoryHandle) {
       if (handle.kind === 'file') {
@@ -252,6 +301,7 @@ export class SyncAccessHandleVFS extends VFS.Base {
     // Verify the digest.
     const computedDigest = this.#computeDigest(encodedPath);
     if (fileDigest.every((value, i) => value === computedDigest[i])) {
+      // Good digest. Decode the null-terminated path string.
       const pathBytes = encodedPath.findIndex(value => value === 0);
       return new TextDecoder().decode(encodedPath.subarray(0, pathBytes));
     } else {
@@ -266,16 +316,21 @@ export class SyncAccessHandleVFS extends VFS.Base {
    * @param {string} path
    */
   #setAssociatedPath(accessHandle, path) {
+    // Convert the path string to UTF-8 and get the digest.
     const encodedPath = new Uint8Array(HEADER_MAX_PATH_SIZE);
     const encodedResult = new TextEncoder().encodeInto(path, encodedPath);
     if (encodedResult.written >= encodedPath.byteLength) {
       throw new Error('path too long');
     }
-
     const digest = this.#computeDigest(encodedPath);
+
+    // Write the OPFS file header.
     accessHandle.write(encodedPath, { at: HEADER_OFFSET_PATH });
     accessHandle.write(digest, { at: HEADER_OFFSET_DIGEST });
+
     if (!path) {
+      // This OPFS file doesn't represent any SQLite file so it doesn't
+      // need to keep any data.
       accessHandle.truncate(HEADER_OFFSET_DATA);
     }
     accessHandle.flush();
@@ -291,6 +346,7 @@ export class SyncAccessHandleVFS extends VFS.Base {
   }
 
   /**
+   * We need a synchronous digest function so can't use WebCrypto.
    * Adapted from https://github.com/bryc/code/blob/master/jshash/experimental/cyrb53.js
    * @param {Uint8Array} corpus 
    * @returns {ArrayBuffer}
@@ -324,6 +380,7 @@ export class SyncAccessHandleVFS extends VFS.Base {
   #deletePath(path) {
     const accessHandle = this.#mapPathToAccessHandle.get(path);
     if (accessHandle) {
+      // Un-associate the SQLite path from the OPFS file.
       this.#setAssociatedPath(accessHandle, '');
       this.#mapPathToAccessHandle.delete(path);
